# specs/001-user-role-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: User Role Management API Contract
  version: 1.0.0
  description: Contract for authentication, role lookup, and role-based user management.
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
  /auth/sign-in:
    post:
      summary: Sign in with username and password
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
      responses:
        '200':
          description: Sign-in successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignInResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials or deleted user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/role:
    get:
      summary: Resolve role from bearer token
      tags: [Authentication]
      responses:
        '200':
          description: Role resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    SignInRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 120
        password:
          type: string
          minLength: 1
    SignInResponse:
      type: object
      required: [token, tokenType, issuedAtUtc, role]
      properties:
        token:
          type: string
        tokenType:
          type: string
          example: Bearer
        issuedAtUtc:
          type: string
          format: date-time
        expiresAtUtc:
          type: string
          format: date-time
          nullable: true
          description: Null by design because automatic token expiry is disabled.
        role:
          type: string
          enum: [user, admin, agent]
    RoleResponse:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [user, admin, agent]
...
    ErrorResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
          example: validation_error
        message:
          type: string
        traceId:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
...